use a client module folder
